+++ CHANGELOG.md
@@ -9,6 +9,7 @@ BREAKING CHANGES
 IMPROVEMENT
 
 - [core] Delete coins with 0 reserves #217
+- [genesis] Add option to export/import state
 - [api] Add ?include_stakes to /candidates endpoint #222
 
 BUG FIXES